chore: use setup i18n plugin for client and server
